requirements.txt
fastapi[all]==0.115.2
uvicorn[standard]==0.32.0
openai==1.52.0
python-dotenv==1.0.1
beanie==1.27.0
pydub==0.25.1
python-jose 
passlib[bcrypt]
